In order to avoid circular imports, adopt a convention that follows astropy.  The main functions are put into a subpackage called scripts.  These files contain the main and parse functions, and then import the desispec.<blah> package for algorithms and the desispec.io.<blah> package for read/write.  Previously, the algorithm and I/O packages were importing each other, which caused an error.  Both the bootcalib and fiberflat packages and their unit tests are updated to this new organization.
