+++ py/desispec/scripts/__init__.py
@@ -1,0 +1,14 @@
+#
+# See top-level LICENSE.rst file for Copyright information
+#
+# -*- coding: utf-8 -*-
+"""
+desispec.scripts
+====================
+
+Main functions and commandline parsing.
+"""
+
+# help with 2to3 support
+from __future__ import absolute_import, division
+
